Set a maximum fps
This is a hack to prevent lock starvation for the physics & light
threads for the BOX2D_LOCK lock.

Sadly the light thread needs to be done on the render thread while
having the BOX2D_LOCK. This means that it is not really possible to
update the physics independently from updating the frame.

A proper fix to this is to use a proper Lock, not synchronize
over an object.